v3 stage 1b: the Xcode-side warnings -- 64-bit casts made explicit, formats fixed
Every non-deprecation warning the CI audit listed outside the C sweep:

- the project passed -fsingle-precision-constant, a GCC flag clang has
  never honoured (removed: constants were always double here);
- 64->32 truncations made explicit where the value is bounded (strlen,
  ftell, recvfrom, CGImage sizes, OpenAL sizes) -- and honest where it
  was not: the timer's epoch base is a time_t now;
- format strings: OSStatus as int, NSInteger as long;
- setFrameInterval: (deprecated, and ambiguous between two classes on an
  id receiver) replaced by preferredFramesPerSecond on a typed pointer;
- CGImageAlphaInfo cast to CGBitmapInfo, as the API wants;
- FS_InitFilesystem is bool and finally returns; a shader-null log line
  that sat after its own return now runs.

Still standing, on purpose: the 511 deprecation sites (OpenGL ES / EAGL /
OpenAL / CoreAudio) -- those are the APIs stage 3 retires, and the launch
image (stage 2 replaces it with a storyboard).

+++ engine/iOS/AQ/AQPlayer.mm
@@ -65,7 +65,7 @@ void AQPlayer::AQBufferCallback(void *					inUserData,
 	UInt32 nPackets = THIS->GetNumPacketsToRead();
 	OSStatus result = AudioFileReadPackets(THIS->GetAudioFileID(), false, &numBytes, inCompleteAQBuffer->mPacketDescriptions, THIS->GetCurrentPacket(), &nPackets, inCompleteAQBuffer->mAudioData);
 	if (result)
-		printf("AudioFileReadPackets failed: %ld", result);
+		printf("AudioFileReadPackets failed: %d", (int)result);
 	if (nPackets > 0) {
 		inCompleteAQBuffer->mAudioDataByteSize = numBytes;		
 		inCompleteAQBuffer->mPacketDescriptionCount = nPackets;
